Start libcompat.a source.
libcompat.a is a static library containing a subset of library functions
from OpenBSD.  These are things that are used by multiple programs and
will work on Linux with just recompiling and minimal edits.  This
library is not intended to be built and installed as a shared library on
the system.