Patch nl.c to get it working on Linux.
A number of things.  The __dead and removal of pledge() is typical.
This program also used the BSD __progname thing, which is not present on
Linux.  Replace the getprogname() call with passing around the result of
basename(argv[0]) which also required changing the usage() function to
accepting a const char *.  There was an errc() call to replace with
err() and the program needed to link with libcompat too.  We also need
to include libgen.h and define _GNU_SOURCE when building so we get the
getline() definition as well as NL_TEXTMAX which is actually pulled in
through __USE_XOPEN which we indirectly get via _GNU_SOURCE.