Get join.c compiling, but it's likely still broken.
My first pass through everything is to get things compiling on Linux
with glibc.  We'll need a test suite to make sure things pass expected
use cases, but that's later.

Aside from removing pledge() and __dead from join.c and linking it with
libcompat, we also have to replace the fgetln() call.  fgetln() is
specific to BSD.  POSIX defines getline() as the standard replacement
which, surprisingly, came from GNU.  The way you use it is a little
different and I expect this to need some changes.