Add a block to configure.ac to set the UID_MAX value.
UID_MAX is compiled in to the id program.  It's not in a header file,
but rather is set by /etc/login.defs.  Whether or not this should be
compiled in is another question, but at least for now I've added a block
to set the UID_MAX value at compile time -but- if you don't pass that it
will just pick up the value from your system if it can.  Ultimately it
defaults to 60000.