Big changes to arch.c to get it working on Linux.
arch.c on OpenBSD has information set at compile time.  The system
headers provide a file called <sys/param.h> which in turn includes
<machine/param.h>.  The latter file is linked to the appropriate
<sys/arch/MACHINE/param.h> file where MACHINE is sgi, alpha, i386,
amd64, mips64, powerpc, and so on.

The arch.c program only uses the MACHINE and MACHINE_ARCH macros from
that param.h file and for each architecture it's just a string matching
the directory name.  So on amd64, MACHINE is "amd64" and MACHINE_ARCH
is also "amd64".  The BSD arch command will output OSNAME.MACHINE as the
default output, but even that is hardcoded to OpenBSD.  We don't really
want the arch program on Linux saying that but we also don't want to
hardcode it to Linux.  So...

Remove the use of <sys/param.h> and the MACHINE* macros.  Instead call
the uname(2) function and fill a struct utsname.  From this pull
.machine for the architecture and .sysname for the operating system.  On
Linux this can be useful if you, say, boot an x86_64 system with an i386
kernel.  The arch command doesn't need to be recompiled because it just
gets the information at runtime.  Likewise, if you change the operating
system name from Linux to Stanley, it will show that as well without
recompiling the program.