Fix chmod.c to build and work on Linux.
Aside from the usual removal of __dead and pledge() calls, also remove
all support for the chflags program.  This doesn't exist on Linux.
Instead we have per-filesystem attr programs.  Remove the man page and
support for this program in the source.

Also remove UID_MAX and GID_MAX with INT_MAX.  This may have to change
in the future.