Update the import-src.sh for rebases.
When a new release of OpenBSD comes out, we'll want the ability to
rebase our source and then manually check things against our git
commits.  We need to copy in .c and .h files at a minimum and see if
there are any changes in the upstream Makefile rule changes.  The entire
process of rebasing will be tedious, so probably not something we do all
the time.