Get stty(1) working on Linux.
This one required a lot of work.  See the top level DIFFERENCES file for
an explanation of the changes you see in the stty code.  For the most
part it was updating the code to use Linux macro names, removing things
that just don't exist in Linux, and then rewriting things that are
handled a different way in Linux.