Add back code that was accidentally deleted from stat.c
This is required for the SHOW_st_{a,m,c}time fallthroughs.  Also fix up
the man page to reflect the changes made.

st_birthtime is per filesystem on Linux and because of that does not
appear as a struct stat member as it does on OpenBSD.  I am not opposed
to adding in support for st_birthtime for supported filesystems, but it
does mean that stat(1) would have to link with any number of additional
filesystem libraries in order to get that information.  Might be better
to leave it out of stat(1) and just have users rely on filesystem tools
for that information.