Use .tv_sec and .tv_nsec in ls similar to stat.
The glibc feature test macros expose different convenience struct member
names in struct stat.  Rather than rely on the BSD names to be exposed
with the right set of feature test macros, change the code over to using
the .tv_sec and .tv_nsec struct members that are always available.