Final fixes for sort to get it compiling and running.
There are likely still bugs (in fact I see some in local output I'm
generating).  Major parts of this patch:

1) Link with OpenSSL or LibreSSL for MD5 functions.  glibc doesn't have
these.  We need MD5 support for sort's random function.  Linking with
the SSL libraries has it use the same basic API as it does on OpenBSD.

2) Add a test to configure.ac for openssl.pc.  Might have to wrap this
in a enable/disable flag so that you can build on systems without
OpenSSL or LibreSSL and still get everything but sort.

3) Remove all the pledge() calls.

4) Use getrandom() instead of arc4random().

5) Use sysconf() to compute the physical memory installed.