Remove the scripts/ subdirectory.
The only thing in here is tac.in, which is a wrapper that runs "tail -r"
because tac is only a GNU coreutils program.  It's cat backwards, which
tail -r can do.  But tac does have some additional options that tail
does not provide, so it's not really a direct emulation of tac.  You
want bsdutils, you don't get tac.