Update compat/ sources with OpenBSD 6.7 code.
Bring in strlcpy() as well.  compat.h carries macros to cover things
defined on OpenBSD but not on Linux, or at least slightly differently
on Linux.  compat.h is also included in to all of the .c files in
compat/ so they can read those definitions.

devname.c has the devname_nodb() function renamed to devname() because
the new devname() function uses the OpenBSD dev.db which we do not
have on Linux.  The nodb variant of the function is what we can use.