First batch of OpenBSD 6.7 src/ commands patched and building.
Trying a new mechanism to allow for easier updating.  The source tree
remains in the same style, but I will keep patches in patches/ that
are applied by the import-src.sh script.

The patches do not contain all of the changes since some changes are
handled by sed from import-src.sh