Port df(1) to Linux
This command is likely the most different from OpenBSD when in comes
to the code.  See the DIFFERENCES file for some details.  The
getmntinfo() function does not exist, so I implemented that locally in
df.c.  I also got rid of the raw_df() function and the ffs_df.c and
ext2fs_df.c files because on Linux we just fall through to what
statvfs() tells us.

There are likely bugs in this command since it is such a different
port.  Please report any issues found.