Begin import and porting of FreeBSD 13.0-RELEASE code
Rather than continuing with OpenBSD porting, move over to FreeBSD for
slightly more featureful commands.  Another advantage is this puts
bsdutils at a better compatibility state with users of MacOS X (or so
I'm told).  I spent a few years porting from OpenBSD here and there
and it was ok, but now it's time to try the FreeBSD code base.