WIP: update docs and other files during FreeBSD import
I'm actively porting from FreeBSD 13.0-RELEASE and this is just the
current state of the tree.  See src/Makefile and the COMMANDS variable
for a list of things I have completed.  No idea when I will finish,
but it is going pretty quickly because most of what I did for OpenBSD
applies to FreeBSD code.