Get ls(1), test(1), and users(1) over to meson.build
ls requires libtinfo, so add a dependency for that.  test(1) can't be
called 'test' because that's a reserved meson string.  Build it as
xtest but install it as 'test'.  users(1) is written in C++, so add
all of that to the top level meson.build file so we can build for C or
C++.  I really don't like that users(1) is in C++, so I may actually
change that later.