Port df(1) to Linux
This is a more complicated command to port.  I used the same trick I
did on OpenBSD which is to implement a local getmntinfo() and
freemntinfo() function.  FreeBSD uses libxo for output, so that is now
a runtime requirement.

NOTE: This is a work in progress and I am still testing it and
revising the port.