Implement the f_selected mechanism for fs selection in df(1)
Because we have to implement getmntinfo() directly, we collect
everything via getmntent().  df(1) on FreeBSD is more flexible so it
can restrict getmntinfo() to the filesystems specified.  The
workaround here is to collect everything in one pass and then mark
f_selected based on whether or not the user specified one or more
mount points on the command line.