ls(1): fall back to numeric IDs in long form if needed
This better matches the behavior of other ls(1) implementations
as it will not just fail but fall back to actual numeric IDs if
the real names are not known. This can happen e.g. in a chroot
or a namespace or any place where you have files or directories
with UIDs/GIDs that do not match what is in the local database.