First pass at porting timeout(1)
So it seems to work, but for some processes it seems to not kill
things cleanly.  That may be expected and I have no compared what
happens using timeout from GNU coreutils.  I really have only seen it
with interactive programs with ncurses interfaces.  But for things
like dd, tar, cp, and other types of operations I have it working.  I
set a specific runtime and it kills the process when the timer hits.